// sanity.types.ts
export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};
// ...
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// ...
export type CASE_STUDY_CATEGORIES_QUERYResult = Array<{
  _id: string;
  title: string | null;
  slug: Slug | null;
}>;
// Variable: CASE_STUDIES_QUERY
// Query: *[_type == "caseStudy" && defined(slug.current)][0...50] | order(publishedAt desc) {    _id,    title,    description,    slug,    mainImage,    publishedAt,    _updatedAt,    featured,    "category": category->title,  }
export type CASE_STUDIES_QUERYResult = Array<{
  _id: string;
  title: string | null;
  description: string | null;
  slug: Slug | null;
  mainImage: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    alt?: string;
    _type: "image";
  } | null;
  publishedAt: string | null;
  _updatedAt: string;
  featured: boolean | null;
  category: string | null;
}>;